examples/web_app_project/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
